Moved buildIndex to separate file
